requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-filter== 23.5
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise
